+++ packages/polaris-viz/src/hooks/index.ts
@@ -29,5 +29,8 @@ export {
   getSeriesColors,
   usePrevious,
 } from '@shopify/polaris-viz-core';
-export {useWatchActiveSeries} from './useWatchActiveSeries';
+export {
+  useWatchActiveSeries,
+  setActiveSeriesListener,
+} from './useWatchActiveSeries';
 export {useExternalHideEvents, setHiddenItems} from './ExternalEvents';